requirements.txt
openai
langchain
chromadb
streamlit
python-dotenv